secret_voldemort/Backend/requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
pony==0.7.13
pytest==5.3.5
requests
python-multipart
pydantic[email]
python-jose[cryptography]
passlib[Bcrypt]
email-validator

